Refactor metadata controller into a package
Split the single-file metadata controller into a folder-based package so it can
carry its own strings.json (core.metadata namespace) for server-owned
name/description localization, matching the other folder-based core controllers.

- metadata.py -> metadata/ package: thin __init__.py re-exporting
  MetaDataController, controller.py for the controller class, and cohesive
  ImageProxyMixin (images.py), RadioArtworkMixin (radio.py) and
  MetadataEnrichmentMixin (enrichment.py) mixins, plus helpers.py/constants.py
- add strings.json with the manifest name/description and regenerate
  translations/en.json
- add a README documenting the package architecture
- add unit tests for the radio name-matching and image helpers
- behaviour unchanged; public import path and MetaDataController are identical
