+++ music_assistant/controllers/metadata/__init__.py
@@ -1,0 +1,7 @@
+"""MusicAssistant MetaDataController package."""
+
+from __future__ import annotations
+
+from .controller import MetaDataController
+
+__all__ = ["MetaDataController"]
